+++ packages/fighting-design/swap/index.ts
@@ -4,6 +4,7 @@ import { install } from '../_utils'
 
 export const FSwap = install(Swap)
 
+/** swap 组件实例类型 */
 export type SwapInstance = InstanceType<typeof Swap>
 
 export * from './src/interface.d'