+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
 - `f-progress` 组件 `type` 新增 `info` 和 `default` 配置项
+- `f-swap` 组件 `type` 废除 `default` 配置项
 
 ## 0.23.1 (2023-02-16)