+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+## 0.24.0 (2023-02-17)
+
 - `f-progress` 组件 `type` 新增 `info` 和 `default` 配置项
 - `f-swap` 组件 `type` 废除 `default` 配置项
 - `f-link` 组件 `state` 配置项的 `bag` 可选值改为 `background`